repo reorg: move network/http ratelimit filters
Core ratelimit functionality has been left in the server as it is likely that
will be used for other things in the future (and has a null implementation when
not configured).
