+++ source/common/http/filter/BUILD
@@ -119,35 +119,3 @@ envoy_cc_library(
         "@envoy_api//envoy/config/filter/http/squash/v2:squash_cc",
     ],
 )
-
-envoy_cc_library(
-    name = "ratelimit_lib",
-    srcs = ["ratelimit.cc"],
-    deps = [
-        ":ratelimit_includes",
-        "//include/envoy/http:codes_interface",
-        "//source/common/common:assert_lib",
-        "//source/common/common:empty_string",
-        "//source/common/common:enum_to_int",
-        "//source/common/http:codes_lib",
-        "//source/common/router:config_lib",
-    ],
-)
-
-envoy_cc_library(
-    name = "ratelimit_includes",
-    hdrs = ["ratelimit.h"],
-    deps = [
-        "//include/envoy/http:filter_interface",
-        "//include/envoy/local_info:local_info_interface",
-        "//include/envoy/ratelimit:ratelimit_interface",
-        "//include/envoy/runtime:runtime_interface",
-        "//include/envoy/upstream:cluster_manager_interface",
-        "//source/common/common:assert_lib",
-        "//source/common/http:header_map_lib",
-        "//source/common/json:config_schemas_lib",
-        "//source/common/json:json_loader_lib",
-        "//source/common/json:json_validator_lib",
-        "@envoy_api//envoy/config/filter/http/rate_limit/v2:rate_limit_cc",
-    ],
-)
