+++ source/exe/BUILD
@@ -48,12 +48,10 @@ envoy_cc_library(
         "//source/server/config/http:gzip_lib",
         "//source/server/config/http:ip_tagging_lib",
         "//source/server/config/http:lua_lib",
-        "//source/server/config/http:ratelimit_lib",
         "//source/server/config/http:router_lib",
         "//source/server/config/listener:original_dst_lib",
         "//source/server/config/listener:proxy_protocol_lib",
         "//source/server/config/network:http_connection_manager_lib",
-        "//source/server/config/network:ratelimit_lib",
         "//source/server/config/network:raw_buffer_socket_lib",
         "//source/server/config/network:ssl_socket_lib",
         "//source/server/config/stats:dog_statsd_lib",
