+++ source/extensions/all_extensions.bzl
@@ -5,10 +5,12 @@
 def envoy_all_extensions(repository = ""):
   return [
     repository + "//source/extensions/filters/http/ext_authz:config",
+    repository + "//source/extensions/filters/http/ratelimit:config",
     repository + "//source/extensions/filters/network/client_ssl_auth:config",
     repository + "//source/extensions/filters/network/echo:config",
     repository + "//source/extensions/filters/network/ext_authz:config",
     repository + "//source/extensions/filters/network/mongo_proxy:config",
+    repository + "//source/extensions/filters/network/ratelimit:config",
     repository + "//source/extensions/filters/network/redis_proxy:config",
     repository + "//source/extensions/filters/network/tcp_proxy:config",
   ]
