+++ source/extensions/filters/http/ratelimit/BUILD
@@ -1,6 +1,6 @@
 licenses(["notice"])  # Apache 2
-# External authorization L7 HTTP filter
-# Public docs: TODO(saumoh): Docs needed in https://github.com/envoyproxy/data-plane-api/tree/master/docs/root/configuration/http_filters
+# Ratelimit L7 HTTP filter
+# Public docs: https://github.com/envoyproxy/data-plane-api/blob/master/docs/root/configuration/http_filters/rate_limit_filter.rst
 
 load(
     "//bazel:envoy_build_system.bzl",
@@ -11,9 +11,9 @@ load(
 envoy_package()
 
 envoy_cc_library(
-    name = "ext_authz",
-    srcs = ["ext_authz.cc"],
-    hdrs = ["ext_authz.h"],
+    name = "ratelimit_lib",
+    srcs = ["ratelimit.cc"],
+    hdrs = ["ratelimit.h"],
     deps = [
         "//include/envoy/http:codes_interface",
         "//source/common/common:assert_lib",
@@ -21,8 +21,7 @@ envoy_cc_library(
         "//source/common/common:enum_to_int",
         "//source/common/http:codes_lib",
         "//source/common/router:config_lib",
-        "//source/extensions/filters/common/ext_authz:ext_authz_lib",
-        "@envoy_api//envoy/config/filter/http/ext_authz/v2:ext_authz_cc",
+        "@envoy_api//envoy/config/filter/http/rate_limit/v2:rate_limit_cc",
     ],
 )
 
@@ -31,9 +30,10 @@ envoy_cc_library(
     srcs = ["config.cc"],
     hdrs = ["config.h"],
     deps = [
-        ":ext_authz",
+        ":ratelimit_lib",
         "//include/envoy/registry",
         "//include/envoy/server:filter_config_interface",
+        "//source/common/config:filter_json_lib",
         "//source/common/config:well_known_names",
         "//source/common/protobuf:utility_lib",
     ],
