+++ source/extensions/filters/http/ratelimit/config.h
@@ -1,0 +1,44 @@
+#pragma once
+
+#include <string>
+
+#include "envoy/config/filter/http/rate_limit/v2/rate_limit.pb.h"
+#include "envoy/server/filter_config.h"
+
+#include "common/config/well_known_names.h"
+
+namespace Envoy {
+namespace Extensions {
+namespace HttpFilters {
+namespace RateLimitFilter {
+
+/**
+ * Config registration for the rate limit filter. @see NamedHttpFilterConfigFactory.
+ */
+class RateLimitFilterConfig : public Server::Configuration::NamedHttpFilterConfigFactory {
+public:
+  Server::Configuration::HttpFilterFactoryCb
+  createFilterFactory(const Json::Object& json_config, const std::string&,
+                      Server::Configuration::FactoryContext& context) override;
+
+  Server::Configuration::HttpFilterFactoryCb
+  createFilterFactoryFromProto(const Protobuf::Message& proto_config,
+                               const std::string& stats_prefix,
+                               Server::Configuration::FactoryContext& context) override;
+
+  ProtobufTypes::MessagePtr createEmptyConfigProto() override {
+    return ProtobufTypes::MessagePtr{new envoy::config::filter::http::rate_limit::v2::RateLimit()};
+  }
+
+  std::string name() override { return Config::HttpFilterNames::get().RATE_LIMIT; }
+
+private:
+  Server::Configuration::HttpFilterFactoryCb
+  createFilter(const envoy::config::filter::http::rate_limit::v2::RateLimit& proto_config,
+               const std::string& stats_prefix, Server::Configuration::FactoryContext& context);
+};
+
+} // namespace RateLimitFilter
+} // namespace HttpFilters
+} // namespace Extensions
+} // namespace Envoy
