+++ source/extensions/filters/network/client_ssl_auth/BUILD
@@ -1,4 +1,6 @@
 licenses(["notice"])  # Apache 2
+# Client SSL authorization L4 network filter
+# Public docs: https://github.com/envoyproxy/data-plane-api/blob/master/docs/root/configuration/network_filters/client_ssl_auth_filter.rst
 
 load(
     "//bazel:envoy_build_system.bzl",
