+++ source/extensions/filters/network/echo/BUILD
@@ -1,4 +1,6 @@
 licenses(["notice"])  # Apache 2
+# Echo L4 network filter. This is primarily a simplistic example.
+# Public docs: https://github.com/envoyproxy/data-plane-api/blob/master/docs/root/configuration/network_filters/echo_filter.rst
 
 load(
     "//bazel:envoy_build_system.bzl",
