+++ source/extensions/filters/network/redis_proxy/BUILD
@@ -1,4 +1,7 @@
 licenses(["notice"])  # Apache 2
+# Redis proxy L4 network filter. Implements consistent hashing and observability for large redis
+# clusters.
+# Public docs: https://github.com/envoyproxy/data-plane-api/blob/master/docs/root/configuration/network_filters/redis_proxy_filter.rst
 
 load(
     "//bazel:envoy_build_system.bzl",
