--- source/server/config/http/ratelimit.h
@@ -1,39 +1,0 @@
-#pragma once
-
-#include <string>
-
-#include "envoy/config/filter/http/rate_limit/v2/rate_limit.pb.h"
-#include "envoy/server/filter_config.h"
-
-#include "common/config/well_known_names.h"
-
-namespace Envoy {
-namespace Server {
-namespace Configuration {
-
-/**
- * Config registration for the rate limit filter. @see NamedHttpFilterConfigFactory.
- */
-class RateLimitFilterConfig : public NamedHttpFilterConfigFactory {
-public:
-  HttpFilterFactoryCb createFilterFactory(const Json::Object& json_config, const std::string&,
-                                          FactoryContext& context) override;
-  HttpFilterFactoryCb createFilterFactoryFromProto(const Protobuf::Message& proto_config,
-                                                   const std::string& stats_prefix,
-                                                   FactoryContext& context) override;
-
-  ProtobufTypes::MessagePtr createEmptyConfigProto() override {
-    return ProtobufTypes::MessagePtr{new envoy::config::filter::http::rate_limit::v2::RateLimit()};
-  }
-
-  std::string name() override { return Config::HttpFilterNames::get().RATE_LIMIT; }
-
-private:
-  HttpFilterFactoryCb
-  createFilter(const envoy::config::filter::http::rate_limit::v2::RateLimit& proto_config,
-               const std::string& stats_prefix, FactoryContext& context);
-};
-
-} // namespace Configuration
-} // namespace Server
-} // namespace Envoy
