+++ test/common/http/filter/BUILD
@@ -84,28 +84,6 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
-    name = "ratelimit_test",
-    srcs = ["ratelimit_test.cc"],
-    deps = [
-        "//source/common/buffer:buffer_lib",
-        "//source/common/common:empty_string",
-        "//source/common/config:filter_json_lib",
-        "//source/common/http:headers_lib",
-        "//source/common/http/filter:ratelimit_includes",
-        "//source/common/http/filter:ratelimit_lib",
-        "//source/common/ratelimit:ratelimit_lib",
-        "//test/mocks/http:http_mocks",
-        "//test/mocks/local_info:local_info_mocks",
-        "//test/mocks/ratelimit:ratelimit_mocks",
-        "//test/mocks/runtime:runtime_mocks",
-        "//test/mocks/tracing:tracing_mocks",
-        "//test/mocks/upstream:upstream_mocks",
-        "//test/test_common:utility_lib",
-        "@envoy_api//envoy/config/filter/http/ext_authz/v2:ext_authz_cc",
-    ],
-)
-
-envoy_cc_test(
     name = "squash_filter_test",
     srcs = ["squash_filter_test.cc"],
     deps = [
