+++ test/common/network/filter_manager_impl_test.cc
@@ -3,11 +3,11 @@
 
 #include "common/buffer/buffer_impl.h"
 #include "common/config/filter_json.h"
-#include "common/filter/ratelimit.h"
 #include "common/network/filter_manager_impl.h"
 #include "common/stats/stats_impl.h"
 #include "common/upstream/upstream_impl.h"
 
+#include "extensions/filters/network/ratelimit/ratelimit.h"
 #include "extensions/filters/network/tcp_proxy/tcp_proxy.h"
 
 #include "test/common/upstream/utility.h"
@@ -177,11 +177,12 @@ TEST_F(NetworkFilterManagerTest, RateLimitAndTcpProxy) {
   envoy::config::filter::network::rate_limit::v2::RateLimit proto_config{};
   Config::FilterJson::translateTcpRateLimitFilter(*json_config, proto_config);
 
-  RateLimit::TcpFilter::ConfigSharedPtr rl_config(new RateLimit::TcpFilter::Config(
-      proto_config, factory_context.scope_, factory_context.runtime_loader_));
+  Extensions::NetworkFilters::RateLimitFilter::ConfigSharedPtr rl_config(
+      new Extensions::NetworkFilters::RateLimitFilter::Config(proto_config, factory_context.scope_,
+                                                              factory_context.runtime_loader_));
   RateLimit::MockClient* rl_client = new RateLimit::MockClient();
-  manager.addReadFilter(ReadFilterSharedPtr{
-      new RateLimit::TcpFilter::Instance(rl_config, RateLimit::ClientPtr{rl_client})});
+  manager.addReadFilter(std::make_shared<Extensions::NetworkFilters::RateLimitFilter::Filter>(
+      rl_config, RateLimit::ClientPtr{rl_client}));
 
   envoy::config::filter::network::tcp_proxy::v2::TcpProxy tcp_proxy;
   tcp_proxy.set_stat_prefix("name");
