+++ test/extensions/filters/network/client_ssl_auth/BUILD
@@ -14,14 +14,22 @@ envoy_cc_test(
     data = glob(["test_data/**"]),
     deps = [
         "//source/common/config:filter_json_lib",
-        "//source/extensions/filters/network/client_ssl_auth:config",
+        "//source/extensions/filters/network/client_ssl_auth",
         "//test/mocks/network:network_mocks",
         "//test/mocks/runtime:runtime_mocks",
-        "//test/mocks/server:server_mocks",
         "//test/mocks/ssl:ssl_mocks",
         "//test/mocks/thread_local:thread_local_mocks",
         "//test/mocks/upstream:upstream_mocks",
         "//test/test_common:environment_lib",
         "//test/test_common:utility_lib",
+    ],
+)
+
+envoy_cc_test(
+    name = "config_test",
+    srcs = ["config_test.cc"],
+    deps = [
+        "//source/extensions/filters/network/client_ssl_auth:config",
+        "//test/mocks/server:server_mocks",
     ],
 )
