+++ test/extensions/filters/network/ext_authz/config_test.cc
@@ -1,0 +1,46 @@
+#include "extensions/filters/network/ext_authz/config.h"
+
+#include "test/mocks/server/mocks.h"
+
+#include "gmock/gmock.h"
+#include "gtest/gtest.h"
+
+using testing::Invoke;
+using testing::_;
+
+namespace Envoy {
+namespace Extensions {
+namespace NetworkFilters {
+namespace ExtAuthz {
+
+TEST(NetworkFilterConfigTest, ExtAuthzCorrectProto) {
+  std::string yaml = R"EOF(
+  grpc_service:
+    google_grpc:
+      target_uri: ext_authz_server
+      stat_prefix: google
+  failure_mode_allow: false
+  stat_prefix: name
+)EOF";
+
+  envoy::config::filter::network::ext_authz::v2::ExtAuthz proto_config{};
+  MessageUtil::loadFromYaml(yaml, proto_config);
+
+  NiceMock<Server::Configuration::MockFactoryContext> context;
+  ExtAuthzConfigFactory factory;
+
+  EXPECT_CALL(context.cluster_manager_.async_client_manager_, factoryForGrpcService(_, _))
+      .WillOnce(Invoke([](const envoy::api::v2::core::GrpcService&, Stats::Scope&) {
+        return std::make_unique<NiceMock<Grpc::MockAsyncClientFactory>>();
+      }));
+  Server::Configuration::NetworkFilterFactoryCb cb =
+      factory.createFilterFactoryFromProto(proto_config, context);
+  Network::MockConnection connection;
+  EXPECT_CALL(connection, addReadFilter(_));
+  cb(connection);
+}
+
+} // namespace ExtAuthz
+} // namespace NetworkFilters
+} // namespace Extensions
+} // namespace Envoy
