+++ test/extensions/filters/network/ratelimit/BUILD
@@ -15,11 +15,20 @@ envoy_cc_test(
         "//source/common/buffer:buffer_lib",
         "//source/common/config:filter_json_lib",
         "//source/common/event:dispatcher_lib",
-        "//source/common/filter:ratelimit_lib",
         "//source/common/stats:stats_lib",
+        "//source/extensions/filters/network/ratelimit:ratelimit_lib",
         "//test/mocks/network:network_mocks",
         "//test/mocks/ratelimit:ratelimit_mocks",
         "//test/mocks/runtime:runtime_mocks",
         "//test/mocks/tracing:tracing_mocks",
+    ],
+)
+
+envoy_cc_test(
+    name = "config_test",
+    srcs = ["config_test.cc"],
+    deps = [
+        "//source/extensions/filters/network/ratelimit:config",
+        "//test/mocks/server:server_mocks",
     ],
 )
