+++ test/integration/BUILD
@@ -297,11 +297,9 @@ envoy_cc_test_library(
         "//source/server/config/http:ip_tagging_lib",
         "//source/server/config/http:lightstep_lib",
         "//source/server/config/http:lua_lib",
-        "//source/server/config/http:ratelimit_lib",
         "//source/server/config/http:router_lib",
         "//source/server/config/http:zipkin_lib",
         "//source/server/config/network:http_connection_manager_lib",
-        "//source/server/config/network:ratelimit_lib",
         "//source/server/config/network:raw_buffer_socket_lib",
         "//source/server/config/network:ssl_socket_lib",
         "//source/server/config/stats:statsd_lib",
@@ -410,6 +408,7 @@ envoy_cc_test(
         "//source/common/grpc:codec_lib",
         "//source/common/grpc:common_lib",
         "//source/common/ratelimit:ratelimit_proto",
+        "//source/extensions/filters/http/ratelimit:config",
         "//test/common/grpc:grpc_client_integration_lib",
     ],
 )
