+++ test/server/config/http/BUILD
@@ -16,6 +16,7 @@ envoy_cc_test(
         "//source/common/protobuf",
         "//source/common/protobuf:utility_lib",
         "//source/common/router:router_lib",
+        "//source/extensions/filters/http/ratelimit:config",
         "//source/server/config/http:buffer_lib",
         "//source/server/config/http:dynamo_lib",
         "//source/server/config/http:fault_lib",
@@ -24,7 +25,6 @@ envoy_cc_test(
         "//source/server/config/http:grpc_web_lib",
         "//source/server/config/http:ip_tagging_lib",
         "//source/server/config/http:lua_lib",
-        "//source/server/config/http:ratelimit_lib",
         "//source/server/config/http:router_lib",
         "//source/server/config/http:squash_lib",
         "//source/server/config/http:zipkin_lib",
