+++ test/server/config/network/BUILD
@@ -19,11 +19,11 @@ envoy_cc_test(
         "//source/extensions/filters/network/client_ssl_auth:config",
         "//source/extensions/filters/network/ext_authz:config",
         "//source/extensions/filters/network/mongo_proxy:config",
+        "//source/extensions/filters/network/ratelimit:config",
         "//source/extensions/filters/network/redis_proxy:config",
         "//source/extensions/filters/network/tcp_proxy:config",
         "//source/server/config/access_log:file_access_log_lib",
         "//source/server/config/network:http_connection_manager_lib",
-        "//source/server/config/network:ratelimit_lib",
         "//test/mocks/grpc:grpc_mocks",
         "//test/mocks/server:server_mocks",
         "//test/test_common:utility_lib",
