+++ test/server/config/network/config_test.cc
@@ -10,11 +10,11 @@
 
 #include "server/config/access_log/file_access_log.h"
 #include "server/config/network/http_connection_manager.h"
-#include "server/config/network/ratelimit.h"
 
 #include "extensions/filters/network/client_ssl_auth/config.h"
 #include "extensions/filters/network/ext_authz/config.h"
 #include "extensions/filters/network/mongo_proxy/config.h"
+#include "extensions/filters/network/ratelimit/config.h"
 #include "extensions/filters/network/redis_proxy/config.h"
 #include "extensions/filters/network/tcp_proxy/config.h"
 
@@ -34,6 +34,7 @@ namespace Server {
 namespace Configuration {
 
 // Negative test for protoc-gen-validate constraints.
+// TODO(mattklein123): Break this test apart into per extension tests.
 TEST(NetworkFilterConfigTest, ValidateFail) {
   NiceMock<MockFactoryContext> context;
 
@@ -43,7 +44,7 @@ TEST(NetworkFilterConfigTest, ValidateFail) {
   envoy::config::filter::network::http_connection_manager::v2::HttpConnectionManager hcm_proto;
   Extensions::NetworkFilters::MongoProxy::MongoProxyFilterConfigFactory mongo_factory;
   envoy::config::filter::network::mongo_proxy::v2::MongoProxy mongo_proto;
-  RateLimitConfigFactory rate_limit_factory;
+  Extensions::NetworkFilters::RateLimitFilter::RateLimitConfigFactory rate_limit_factory;
   envoy::config::filter::network::rate_limit::v2::RateLimit rate_limit_proto;
   Extensions::NetworkFilters::RedisProxy::RedisProxyFilterConfigFactory redis_factory;
   envoy::config::filter::network::redis_proxy::v2::RedisProxy redis_proto;
@@ -70,72 +71,6 @@ TEST(NetworkFilterConfigTest, ValidateFail) {
   EXPECT_THROW(FileAccessLogFactory().createAccessLogInstance(
                    envoy::config::filter::accesslog::v2::FileAccessLog(), nullptr, context),
                ProtoValidationException);
-}
-
-TEST(NetworkFilterConfigTest, RatelimitCorrectJson) {
-  std::string json_string = R"EOF(
-  {
-    "stat_prefix": "my_stat_prefix",
-    "domain" : "fake_domain",
-    "descriptors": [[{ "key" : "my_key",  "value" : "my_value" }]],
-    "timeout_ms": 1337
-  }
-  )EOF";
-
-  Json::ObjectSharedPtr json_config = Json::Factory::loadFromString(json_string);
-  NiceMock<MockFactoryContext> context;
-  RateLimitConfigFactory factory;
-  NetworkFilterFactoryCb cb = factory.createFilterFactory(*json_config, context);
-  Network::MockConnection connection;
-  EXPECT_CALL(connection, addReadFilter(_));
-  cb(connection);
-}
-
-TEST(NetworkFilterConfigTest, RatelimitCorrectProto) {
-  std::string json_string = R"EOF(
-  {
-    "stat_prefix": "my_stat_prefix",
-    "domain" : "fake_domain",
-    "descriptors": [[{ "key" : "my_key",  "value" : "my_value" }]],
-    "timeout_ms": 1337
-  }
-  )EOF";
-
-  Json::ObjectSharedPtr json_config = Json::Factory::loadFromString(json_string);
-  envoy::config::filter::network::rate_limit::v2::RateLimit proto_config{};
-  Config::FilterJson::translateTcpRateLimitFilter(*json_config, proto_config);
-
-  NiceMock<MockFactoryContext> context;
-  RateLimitConfigFactory factory;
-  NetworkFilterFactoryCb cb = factory.createFilterFactoryFromProto(proto_config, context);
-  Network::MockConnection connection;
-  EXPECT_CALL(connection, addReadFilter(_));
-  cb(connection);
-}
-
-TEST(NetworkFilterConfigTest, RatelimitEmptyProto) {
-  std::string json_string = R"EOF(
-  {
-    "stat_prefix": "my_stat_prefix",
-    "domain" : "fake_domain",
-    "descriptors": [[{ "key" : "my_key",  "value" : "my_value" }]],
-    "timeout_ms": 1337
-  }
-  )EOF";
-
-  Json::ObjectSharedPtr json_config = Json::Factory::loadFromString(json_string);
-
-  NiceMock<MockFactoryContext> context;
-  RateLimitConfigFactory factory;
-  envoy::config::filter::network::rate_limit::v2::RateLimit proto_config =
-      *dynamic_cast<envoy::config::filter::network::rate_limit::v2::RateLimit*>(
-          factory.createEmptyConfigProto().get());
-  Config::FilterJson::translateTcpRateLimitFilter(*json_config, proto_config);
-
-  NetworkFilterFactoryCb cb = factory.createFilterFactoryFromProto(proto_config, context);
-  Network::MockConnection connection;
-  EXPECT_CALL(connection, addReadFilter(_));
-  cb(connection);
 }
 
 TEST(NetworkFilterConfigTest, BadHttpConnectionMangerConfig) {
